Refact: WeatherListViewController의 날씨 리스트를 가져오는 부분을 UseCase로 분리
VC가 많은 일을 하므로 앞으로 VC는 View에 관한 데이터 처리만을 담당하도록 하기 위해서 UseCase를 도입했습니다.
NSDataAsset을 사용하면 UseCase에서 UIKit을 import해야하기때문에 json파일을 Assets밖으로 빼고 Bundle에서 가져오도록 변경했습니다.
