--- WeatherForecast/WeatherForecast/App/Assets.xcassets/weather.dataset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "data" : [
-    {
-      "filename" : "weather.json",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
